+++ CHANGELOG.md
@@ -143,6 +143,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Conditional configuration now supports `--when.commands` to change configuration
   based on subcommand.
 
+* `jj bookmark forget` now accepts a `--untrack` flag to untrack remote
+  bookmarks instead of forgetting them.
+
 ### Fixed bugs
 
 * `jj git fetch` with multiple remotes will now fetch from all remotes before